feat: convert sources.js to sources.ts
